Stacked cards visually defined
* In order to implement this the left style property has to be
calculated for each stacked card and the margin property has to be
calculated for the first sibling (a tag) after the stacked cards.
* Note: this could have been implemented entirely in the CSS file if the
calc() function supported the use of counter values. At present it does
not, which is a very sad circumstance given how elegant this would have
been to implement otherwise.